+++ moshi-proguard-rule-gen/src/test/kotlin/dev/zacsweers/moshix/proguardgen/MoshiProguardGenSymbolProcessorTest.kt
@@ -97,6 +97,65 @@ class MoshiProguardGenSymbolProcessorTest(private val useKSP2: Boolean) {
   }
 
   @Test
+  fun ensureNestedSealedClassesAreAdded() {
+    val source =
+      kotlin(
+        "BaseType.kt",
+        """
+      package test
+      import com.squareup.moshi.Json
+      import com.squareup.moshi.JsonClass
+      import dev.zacsweers.moshix.sealed.annotations.TypeLabel
+      import dev.zacsweers.moshix.sealed.annotations.NestedSealed
+      import dev.zacsweers.moshix.sealed.annotations.DefaultObject
+
+      @JsonClass(generateAdapter = true, generator = "sealed:type")
+        sealed class Message {
+
+            @TypeLabel("success")
+            @JsonClass(generateAdapter = true)
+            data class Success(
+                @Json(name = "value")
+                val value: String
+            ) : Message()
+
+            @DefaultObject
+            object Unknown : Message()
+        }
+    """,
+      )
+
+    val compilation = prepareCompilation(source)
+    val result = compilation.compile()
+    assertThat(result.exitCode).isEqualTo(ExitCode.OK)
+    val generatedSourcesDir = compilation.kspSourcesDir
+    val proguardFiles = generatedSourcesDir.walkTopDown().filter { it.extension == "pro" }.toList()
+    check(proguardFiles.isNotEmpty())
+    for (generatedFile in proguardFiles) {
+      generatedFile.assertInCorrectPath()
+      when (generatedFile.nameWithoutExtension) {
+        "moshi-test.Message" ->
+          assertThat(generatedFile.readText().trimIndent())
+            .isEqualTo(
+              """
+                  # Conditionally keep this adapter for every possible nested subtype that uses it.
+                  -if class test.Message.Success
+                  -keep class test.MessageJsonAdapter {
+                      public <init>(com.squareup.moshi.Moshi);
+                  }
+                  -if class test.Message.Unknown
+                  -keep class test.MessageJsonAdapter {
+                      public <init>(com.squareup.moshi.Moshi);
+                  }
+                """
+                .trimIndent()
+            )
+        else -> error("Unrecognized proguard file: $generatedFile")
+      }
+    }
+  }
+
+  @Test
   fun disableProguardGeneration() {
     val source =
       kotlin(