+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 - Improve moshi-sealed KSP error messages.
 - Fix fallback adapter support in IR code gen not recognizing Moshi parameters to primary constructors.
 - Check for same subtypes before erroring on duplicate labels in moshi-sealed IR.
+- Fix proguard rule gen not capturing non-sealed subtypes.
 
 0.27.0
 ------